Parameterize Vagrantfile values: RAM, Storage, Network
For more flexibility we now use the first lines on top of the Vagrantfile
to perform customizations. This might allow using the tooling to build
multiple clusters on the same host. Also experimented with lower RAM
requirements (3 GiB) with success so lowering that is an option
for those of you who do not have access to well-equipped hardware.

Signed-off-by: hseipp <20093686+hseipp@users.noreply.github.com>